Create NeuralNetwork.R
The main functions for usage are linearRegressionBGD, linearRegressionSGD, logisticRegressionBGD, logisticRegressionSGD.
You must supply the (X) training data in the format m*n, where m is the number of training examples and n is the number of parameters (not including bias parameters). (y) is the m*1 matrix of target values. (a) is the learning rate/step size for the gradient ascent.